fix more errors

+++ src/vppinfra/elf.h
@@ -966,7 +966,7 @@ elf_get_section_contents (elf_main_t * em,
   result = 0;
   if (vec_len (s->contents) > 0)
     {
-      vec_attr_t va = { .elt_sz = (u32) elt_size };
+      vec_attr_t va = { .elt_sz = elt_size };
       /* Make vector copy of contents with given element size. */
       result =
 	_vec_realloc_internal (result, vec_len (s->contents) / elt_size, &va);
